requirements.txt
pandas>=2.0
openpyxl>=3.1
spacy>=3.8
nltk>=3.9
numpy>=1.24
scipy>=1.10
python-dotenv>=1.0
openai>=1.0
